% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{axes2vectors}
\alias{axes2vectors}
\title{Convert accelerometry to forces}
\usage{
axes2vectors(
  x = NULL,
  path,
  lista = TRUE,
  Time.zone = "Australia/Melbourne",
  ...
)
}
\arguments{
\item{x}{An data.table object with the accelerometer data from the ExpoApp data.It is generated using import_expoapp function.}

\item{path}{It is the path directory to the accelerometer folder of ExpoApp data.}

\item{lista}{Logical variable (TRUE,FALSE). If it is TRUE the function returns two elements the raw data and the 10 second forces dataset. If it false return only the 10 second resolution dataset with forces.}

\item{Time.zone}{Character variable with the time zone information.}

\item{...}{optional arguments of function.}
}
\value{
If lista is TRUE, it returns a list of two data.tables. The first is the raw data (3 axes accelerations) and the second is processed (2 vector forces).
If lista is FALSE, it returns a data.table with the 2 vector forces.
}
\description{
It is the function to convert the x,y,z axes accelerometry measures to 2 vector forces (Vertical & Horizontal).
}
