% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{completeness_expoapp}
\alias{completeness_expoapp}
\title{Evaluation of Data Completeness of an ExpoApp session.}
\usage{
completeness_expoapp(aux, ...)
}
\arguments{
\item{aux}{The data.table with ExpoApp information at minute resolution.}

\item{...}{optional arguments of table function.}
}
\value{
value
}
\description{
It is the function to assess the number of minutes with accelerometer but no location. The off printed identify when the smartphone was turn off.
}
