% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{decrypt.list}
\alias{decrypt.list}
\title{Decrypt a Geolocation ExpoApp Files List}
\usage{
decrypt.list(gps_dir = NULL, output_dir = NULL, SensorLab = NULL, ...)
}
\arguments{
\item{gps_dir}{The path where the encrypted ExpoApp GPS files are stored.}

\item{output_dir}{The path where we want to store the desencrypted ExpoApp gps data.}

\item{SensorLab}{The path to the SensorLab folder with the jar file.}

\item{...}{optional arguments of function.}
}
\value{
value
}
\description{
It is a function to decrypt at once all ExpoApp gps files stored into the same folder.
}
