% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{decrypt}
\alias{decrypt}
\title{Decrypt Geolocation of ExpoApp}
\usage{
decrypt(file = NULL, output_dir = NULL, SensorLab = SensorLab, ...)
}
\arguments{
\item{file}{The path to the encrypted ExpoApp GPS data is stored.}

\item{output_dir}{The path where we want to store the desencrypted ExpoApp gps data.}

\item{SensorLab}{The path to the SensorLab folder with the jar file.}

\item{...}{optional arguments of function.}
}
\value{
value
}
\description{
It is a function to decode the geolocation file of ExpoApp.
}
\examples{
# Using your password and the below link, you can download SensorLab2-1.2.2 tool.
# It contains a jar file and an example dataset.
# Please, unzip and save in your desired path

browseURL("https://cloudstor.aarnet.edu.au/plus/s/5kPnaEyzuRB4cpH")
Lab_folder <-"C:/Users/ddonaire/Documents/SensorLab2-1.2.2"
gps_file <- system.file("extdata", "ExpoApp.GPS.IDddg.csv", package = "ExpoAppRtools")
decrypt(file = gps_file, SensorLab = Lab_folder)

}
