% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt.R
\name{decrypt_expoapp}
\alias{decrypt_expoapp}
\title{Decrypt a Geolocation ExpoApp File}
\usage{
decrypt_expoapp(file = NULL, output_dir = NULL, SensorLab = SensorLab, ...)
}
\arguments{
\item{file}{The path to the encrypted ExpoApp GPS data is stored.}

\item{output_dir}{The path where we want to store the desencrypted ExpoApp gps data.}

\item{SensorLab}{The path to the SensorLab folder with the jar file.}

\item{...}{optional arguments of function.}
}
\value{
The function returns the path to where decrypted file is stored. This function saves a decrypted file of the ExpoApp gps data.
}
\description{
It is a function to decode the geolocation file of ExpoApp.
}
\examples{
# ExpoApp geolocation information is encrypted to ensure the confidentiality of participants
# in case they lose the pheno. Using your password and the below link, you can 
# download SensorLab2-1.2.2 tool. It contains a jar file, a decrypt key and example datasets.
# Please, download, unzip and save SensorLab2-1.2.2 into your desired path.

browseURL("https://cloudstor.aarnet.edu.au/plus/s/5kPnaEyzuRB4cpH")
Lab_folder <-"C:/Users/ddonaire/Documents/SensorLab2-1.2.2"
encryted_file <- file.path(Lab_folder,"ExpoApp.GPS.IDddg.csv")
output_dir <-"C:/Users/ddonaire/Documents"

decrypt_expoapp(file = encrypted_file, SensorLab = Lab_folder, output_dir=output_dir)

}
