% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt.R
\name{decrypt_expoapp_list}
\alias{decrypt_expoapp_list}
\title{Decrypt a Geolocation ExpoApp Files List}
\usage{
decrypt_expoapp_list(gps_dir = NULL, output_dir = NULL, SensorLab = NULL, ...)
}
\arguments{
\item{gps_dir}{The path where the encrypted ExpoApp GPS files are stored.}

\item{output_dir}{The path where we want to store the desencrypted ExpoApp gps data.}

\item{SensorLab}{The path to the SensorLab folder with the jar file.}

\item{...}{optional arguments of function.}
}
\value{
The function returns the path to where the decrypted files are stored. 
This function decrypts all ExpoApp gps files stored in one folder and saves the decrypted file at the output_dir.
}
\description{
It is a function to decrypt at once all ExpoApp gps files stored into the same folder.
}
