% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{Expo_resum}
\alias{Expo_resum}
\title{Expo_resum}
\usage{
Expo_resum(
  ExpoApp = NULL,
  acce_lista = FALSE,
  Time.zone = "Australia/Melbourne",
  output_dir = getwd(),
  ...
)
}
\arguments{
\item{ExpoApp}{It is the ExpoApp RData object}

\item{acce_lista}{A logical variable (TRUE/FALSE) indicating if the accelerometer output should have raw and vectors or only vectors.}

\item{Time.zone}{The time zone of the study area.}

\item{output_dir}{The folder where we want to store the 10 seconds and 1 minute simplified ExpoApp data.}

\item{...}{optional arguments of function.}
}
\value{
value
}
\description{
It is the function to generate a 10 seconds and 1 minute simplified ExpoApp data.
}
\examples{
# Using your password and the below link, you can download SensorLab2-1.2.2 tool.
# It contains a jar file and an example dataset.
# Please, unzip and save it into your desired path.

browseURL("https://cloudstor.aarnet.edu.au/plus/s/5kPnaEyzuRB4cpH")
Lab_folder <-"C:/Users/ddonaire/Documents/SensorLab2-1.2.2"
load(file.path(Lab_folder,"ExpoApp.IDddg.RData"))
ls()
result <- Expo_resum(ExpoApp=expoapp,output_dir=getwd())
sapply(result,class)
}
