% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{expoapp_completness}
\alias{expoapp_completness}
\title{expoapp_completness}
\usage{
expoapp_completness(aux, ...)
}
\arguments{
\item{aux}{The data.table with ExpoApp information at minute resolution.}

\item{...}{optional arguments of table function.}
}
\value{
value
}
\description{
It is the function to assess the number of minutes without location or turn off during monitoring.
}
