% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{gis.expo}
\alias{gis.expo}
\title{gis.expo}
\usage{
gis.expo(
  Build = NULL,
  Decrypted = NULL,
  EPSG_code = 25832,
  Buffer = 150,
  Time.zone = "Europe/Rome",
  ...
)
}
\arguments{
\item{Build}{This the function to clustering. It also enrich data with information about green spaces from OpenStreetMap.}

\item{Decrypted}{This is the path to the desencryted gps data from ExpoApp.}

\item{EPSG_code}{This is the desired projected coordinate reference system of the study area.}

\item{Buffer}{This is the desired minimum buffer used in the clustering algorithm.}

\item{Time.zone}{The time zone of the study area.}

\item{...}{optional arguments of function.}
}
\value{
value It is a sqlite database with the gps data from ExpoApp cleaned and enriched with information about green spaces.
}
\description{
It is the function to clustering the geolocation at places
}
