% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{import_expoapp}
\alias{import_expoapp}
\title{Untar, Decrypt, Save ExpoApp data.}
\usage{
import_expoapp(
  file = NULL,
  SensorLab = NULL,
  Build = NULL,
  EPSG_code = 25832,
  Buffer = 150,
  Time.zone = "Europe/Rome",
  Clustering = FALSE,
  save_RData = TRUE,
  save_untar = TRUE,
  ...
)
}
\arguments{
\item{file}{The path to the encrypted ExpoApp GPS data is stored.}

\item{SensorLab}{The path to the SensorLab folder with the jar file.}

\item{Build}{This the function to clustering. It also enrich data with information about green spaces from OpenStreetMap.}

\item{EPSG_code}{This is the desired projected coordinate reference system of the study area.}

\item{Buffer}{This is the desired minimum buffer used in the clustering algorithm.}

\item{Time.zone}{The time zone of the study area.}

\item{Clustering}{A logical variable (TRUE/FALSE) indicating if applying yes/no the clustering algorithm.}

\item{save_RData}{A logical variable (TRUE/FALSE) indicating if we want to save the RData file of ExpoApp data.}

\item{save_untar}{A logical variable (TRUE/FALSE) indicating if we want to save the untar file of ExpoApp data.}

\item{...}{optional arguments to FUN.}
}
\value{
value
}
\description{
This function untars ExpoApp data, decrypts geolocation ExpoApp data, and saves the untared ExpoApp data and the ExpoApp data in RData format.
}
\examples{
# Using your password and the below link, you can download SensorLab2-1.2.2 tool.
# It contains a jar file and an example dataset.
# Please, unzip and save it into your desired path.

browseURL("https://cloudstor.aarnet.edu.au/plus/s/5kPnaEyzuRB4cpH")
Lab_folder <-"C:/Users/ddonaire/Documents/SensorLab2-1.2.2"
targz_file <- file.path(Lab_folder,"ExpoApp.IDddg.tar.gz")

expoapp <- import_expoapp(file = targz_file, SensorLab = Lab_folder,
                        Time.zone = "Australia/Melbourne", 
                        Clustering = FALSE, save_RData = FALSE, save_untar=FALSE)
sapply(expoapp,head)

gps_sf <- sf::st_as_sf(expoapp$gps,coords=c("LONGITUDE","LATITUDE"),crs=4326)
mapview::mapview(gps_sf)
# see Expoapp_resum to generate the 10 seconds and 1 minute simplified Expoapp files.
}
