% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{mobile.gps.list}
\alias{mobile.gps.list}
\title{mobile.gps.list is the function to convert all clustering sqlite files to RData files.}
\usage{
mobile.gps.list(processed = NULL, ...)
}
\arguments{
\item{processed}{It is the folder with all stored sqlite databases generated from the gps.expo.}

\item{...}{optional arguments of function.}
}
\value{
value
}
\description{
mobile.gps.list is the function to convert all clustering sqlite files to RData files.
}
