% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{mobile.gps}
\alias{mobile.gps}
\title{mobile.gps}
\usage{
mobile.gps(x, ...)
}
\arguments{
\item{x}{It is the sqlite file database generated from the gis.expo function.}

\item{...}{optional arguments of function.}
}
\value{
value
}
\description{
It is the function to simplify and convert the sqlite obtained from clustering function to RData.
}
