% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_expoapp.R
\name{print_expoapp}
\alias{print_expoapp}
\title{Write Html of ExpoApp Quality Data Analysis Report}
\usage{
print_expoapp(
  result,
  output_dir = NULL,
  open_html = TRUE,
  save_html = TRUE,
  ...
)
}
\arguments{
\item{result}{It is a list object with the times, settings, notes, gps_plot, pa_plot, and nolocation.
Times is a data.table with the information of the quality of the data.
Settings are the software details of the phone and settings of the ExpoApp session.
Notes is the logcat from ExpoApp with all the infromation about battery of the phone.
Gps_plot is a mapview object with the gps information from ExpoApp.
PA_plot is a ggplot object with the METs time-series from ExpoApp.
Nolocation is the temporal completeness of the geolocation of the Expoapp session.}

\item{output_dir}{The folder where we want to store the html file with the quality analysis of ExpoApp session.}

\item{open_html}{A logical variable (TRUE/FALSE) indicating if we want to open the Quality Analysis Report in the browser.}

\item{save_html}{A logical variable (TRUE/FALSE) indicating if we want to save the Quality Analysis Report.}

\item{...}{optional arguments to the function.}
}
\value{
value
}
\description{
It is the function to generate the html with the data analysis of the ExpoApp data.
}
