% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{read_gps_expoapp}
\alias{read_gps_expoapp}
\title{Read a Decrypted Geolocation ExpoApp File}
\usage{
read_gps_expoapp(file = NULL, ...)
}
\arguments{
\item{file}{It is a string variable with the path to the decripted ExpoApp GPS data.}

\item{...}{optional arguments of function.}
}
\value{
value
}
\description{
It is a function to read gps data from ExpoApp.
}
\examples{
# Using your password and the below link, you can download SensorLab2-1.2.2 tool.
# It contains a jar file and an example dataset.
# Please, unzip and save it into your desired path.

browseURL("https://cloudstor.aarnet.edu.au/plus/s/5kPnaEyzuRB4cpH")
Lab_folder <-"C:/Users/ddonaire/Documents/SensorLab2-1.2.2"
gps_file <- file.path(Lab_folder,"ExpoApp.GPS.IDddg_decrypted.csv")

gps_data <- read_gps_expoapp(file = gps_file)
head(gps_data)
gps_sf <- sf::st_as_sf(gps_data,coords=c("LONGITUDE","LATITUDE"),crs=4326)
mapview::mapview(gps_sf)

}
