% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{table2frame}
\alias{table2frame}
\title{table2frame is the function to convert tables to data.frames.}
\usage{
table2frame(x, ...)
}
\arguments{
\item{x}{A table object.}

\item{...}{optional arguments of table function.}
}
\value{
A data.frame with the same shape (nrow*ncols) of the table object.
}
\description{
table2frame is the function to convert tables to data.frames.
}
