% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{time2decimal}
\alias{time2decimal}
\title{time2decimal is the function to convert times (HH:MM) to decimals.}
\usage{
time2decimal(date = "2019-12-12 08:20:30", ...)
}
\arguments{
\item{date}{A character o POSIxt variable with format ("yyyy-mm-dd hh:mm:ss").}

\item{...}{optional arguments of table function.}
}
\value{
value
}
\description{
time2decimal is the function to convert times (HH:MM) to decimals.
}
