% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_expoapp.R
\name{import_expoapp}
\alias{import_expoapp}
\title{Untar, Decrypt and Save ExpoApp data.}
\usage{
import_expoapp(
  file = NULL,
  SensorLab = NULL,
  save_RData = TRUE,
  save_untar = TRUE,
  ...
)
}
\arguments{
\item{file}{Character variable with the path to the tar.gz file of ExpoApp data.}

\item{SensorLab}{Character variable with the path to the SensorLab folder. SensorLab folder has to contain the jar and the pem files.}

\item{save_RData}{A logical variable (TRUE/FALSE) indicating if we want to save the RData file of ExpoApp data.}

\item{save_untar}{A logical variable (TRUE/FALSE) indicating if we want to save the untar file of ExpoApp data.}

\item{...}{optional arguments to FUN.}
}
\value{
value
}
\description{
This function untars ExpoApp data, decrypts ExpoApp geolocation file, and saves the ExpoApp untared and in RData format.
}
\examples{
# Using your password and the below link, you can download SensorLab2-1.2.2 tool.
# It contains a jar file and an example dataset.
# Please, unzip and save it into your desired path.

browseURL("https://cloudstor.aarnet.edu.au/plus/s/5kPnaEyzuRB4cpH")
Lab_folder <-"C:/Users/ddonaire/Documents/SensorLab2-1.2.2"
targz_file <- file.path(Lab_folder,"ExpoApp.IDddg.tar.gz")

expoapp <- import_expoapp(file = targz_file, SensorLab = Lab_folder,
                        save_RData = FALSE, save_untar=FALSE)
sapply(expoapp,head)

gps_sf <- sf::st_as_sf(expoapp$gps,coords=c("LONGITUDE","LATITUDE"),crs=4326)
mapview::mapview(gps_sf)
# see Expoapp_resum to generate the 10 seconds and 1 minute simplified Expoapp files.
}
