% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{pa.acti2}
\alias{pa.acti2}
\title{Convert Forces to ActiGraph Counts, METs and Estimates Wearing Time.}
\usage{
pa.acti2(ace, x, plot = FALSE, population = "adults", ...)
}
\arguments{
\item{ace}{A data.table with the accelerometer data from ExpoApp.}

\item{x}{The path, including file name, where we want to save the time-series of METs}

\item{plot}{A logical variable (TRUE/FALSE) indicating png plot with the time-series of METs.}

\item{population}{A character variable ("adults","childhood") indicating if participant was an adult or kid.}

\item{...}{optional arguments of table function.}
}
\value{
value
}
\description{
Convert Forces to ActiGraph Counts, METs and Estimates Wearing Time.
}
