% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_mobile_Tools.R
\name{reduce_expoapp}
\alias{reduce_expoapp}
\title{Generate ExpoApp_Totals, ExpoApp_min ,and Quality Analysis Report}
\usage{
reduce_expoapp(
  ExpoApp = NULL,
  Time.zone = "Australia/Melbourne",
  output_dir = getwd(),
  ...
)
}
\arguments{
\item{ExpoApp}{It is the ExpoApp RData object}

\item{Time.zone}{The time zone of the study area.}

\item{output_dir}{The folder where we want to store the 10 seconds and 1 minute simplified ExpoApp data.}

\item{...}{optional arguments of function.}

\item{acce_lista}{A logical variable (TRUE/FALSE) indicating if the accelerometer output should have raw and vectors or only vectors.}
}
\value{
value
}
\description{
It is the function to generate a 10 seconds and 1 minute simplified ExpoApp data.
}
\examples{
# ExpoApp geolocation information is encrypted to ensure the confidentiality of participants in case they lose the pheno.
# Using your password and the below link, you can download SensorLab2-1.2.2 tool.
# It contains a decrypt key and example datasets.
# Please, download, unzip and save SensorLab2-1.2.2 into your desired path.

browseURL("https://cloudstor.aarnet.edu.au/plus/s/5kPnaEyzuRB4cpH")
Lab_folder <-"C:/Users/ddonaire/Documents/SensorLab2-1.2.2"
load(file.path(Lab_folder,"ExpoApp.IDddg.RData"))
ls()
result <- reduce_expoapp(ExpoApp=expoapp,output_dir=getwd())
sapply(result,class)
}
